+++ app/src/main/java/cl/inacap/tdis08/sapo/captivemonitor/model/Range.java
@@ -3,19 +3,19 @@ package cl.inacap.tdis08.sapo.captivemonitor.model;
 import java.io.Serializable;
 
 public class Range implements Serializable {
-    private int min;
-    private int max;
+    private Double min;
+    private Double max;
 
-    public Range(int min, int max) {
+    public Range(Double min, Double max) {
         this.min = min;
         this.max = max;
     }
 
-    public int getMin() {
+    public Double getMin() {
         return min;
     }
 
-    public int getMax() {
+    public Double getMax() {
         return max;
     }
 }
